parser: move into stmt.rs
